Fix usage of include_relative in the _layout directory.
include_relative doesn't work well in the _layout directory as it is
relative to the rendered page, not the template. Hence if pages are
being built in subdirectories of the website, the include_relative
directive fails to locate the includes.

Change these uses of the include_relative tag to include, and move the
corresponding files to the _includes directory.